merge:delete old db code
