+++ android/gradle.properties
@@ -42,3 +42,5 @@ newArchEnabled=false
 # Use this property to enable or disable the Hermes JS engine.
 # If set to false, you will be using JSC instead.
 hermesEnabled=true
+
+android.disableAutomaticComponentCreation=true
