+++ packages/ui/App.tsx
@@ -1,73 +1,18 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- *
- * @format
- */
-
-import type { PropsWithChildren } from 'react';
-import { SafeAreaView, ScrollView, StatusBar, useColorScheme, View, Text } from 'react-native';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { DebugInstructions, Header, LearnMoreLinks, ReloadInstructions } from 'react-native/Libraries/NewAppScreen';
-import { cryptoTool, setPassword } from 'packages/core/DB/helper';
-
-(async function () {
-  cryptoTool.setGetPasswordMethod(() => '23423');
-  try {
-    const abc = await cryptoTool.encrypt({ qwe: 123 });
-    const qwe = await cryptoTool.decrypt(abc);
-  } catch (e) {
-    console.log(e);
-  }
-})();
-
-type SectionProps = PropsWithChildren<{
-  title: string;
-}>;
-
-function Section({ children, title }: SectionProps): JSX.Element {
-  return (
-    <View className="mt-8 px-2">
-      <Text className="text-2xl text-black dark:text-white">{title}</Text>
-      <Text className="mt-2 text-lg text-black dark:text-white">{children}</Text>
-    </View>
-  );
-}
-
-function Home() {
-  const backgroundStyle = 'bg-neutral-300 dark:bg-slate-900';
-  const isDarkMode = useColorScheme() === 'dark';
-  return (
-    <View style={{ flex: 1 }}>
-      <StatusBar barStyle={isDarkMode ? 'light-content' : 'dark-content'} />
-      <ScrollView contentInsetAdjustmentBehavior="automatic" className={backgroundStyle}>
-        <Header />
-        <View className="bg-white dark:bg-black">
-          <Section title="Step One">
-            1 Edit <Text className="font-bold">App.tsx</Text> to change this screen and then come back to see your edits.
-          </Section>
-          <Section title="See Your Changes">
-            <ReloadInstructions />
-          </Section>
-          <Section title="Debug">
-            <DebugInstructions />
-          </Section>
-          <Section title="Learn More">Read the docs to discover what to do next:</Section>
-          <LearnMoreLinks />
-        </View>
-      </ScrollView>
-    </View>
-  );
-}
+import Welcome from '@pages/Welcome';
+import SetPassword from '@pages/SetPassword';
+import Home from '@pages/Home';
 
 function App(): JSX.Element {
   const Stack = createNativeStackNavigator();
   return (
     <SafeAreaProvider>
       <NavigationContainer>
-        <Stack.Navigator initialRouteName="Home">
+        <Stack.Navigator initialRouteName="Welcome">
+          <Stack.Screen name="Welcome" component={Welcome} />
+          <Stack.Screen name="SetPassword" component={SetPassword} />
           <Stack.Screen name="Home" component={Home} />
         </Stack.Navigator>
       </NavigationContainer>
