+++ tailwind.config.js
@@ -1,4 +1,5 @@
 /** @type {import('tailwindcss').Config} */
+// eslint-disable-next-line no-undef
 module.exports = {
   content: ["./packages/ui/App.{js,jsx,ts,tsx}", "./packages/ui/**/*.{js,jsx,ts,tsx}"],
   theme: {
